loader: move wireguard-specific functionality out to new wireguard.go
