+++ pkg/datapath/loader/loader.go
@@ -4,29 +4,22 @@
 package loader
 
 import (
-	"context"
-	"fmt"
 	"log/slog"
 	"path/filepath"
 	"sync"
 
-	"github.com/cilium/ebpf"
 	"github.com/cilium/hive/cell"
 	"github.com/cilium/hive/job"
 	"github.com/cilium/statedb"
-	"github.com/vishvananda/netlink"
 
 	"github.com/cilium/cilium/pkg/bpf"
-	"github.com/cilium/cilium/pkg/datapath/config"
 	routeReconciler "github.com/cilium/cilium/pkg/datapath/linux/route/reconciler"
 	"github.com/cilium/cilium/pkg/datapath/linux/sysctl"
-	datapathOption "github.com/cilium/cilium/pkg/datapath/option"
 	"github.com/cilium/cilium/pkg/datapath/tables"
 	datapath "github.com/cilium/cilium/pkg/datapath/types"
 	"github.com/cilium/cilium/pkg/defaults"
 	"github.com/cilium/cilium/pkg/endpointstate"
 	"github.com/cilium/cilium/pkg/lock"
-	"github.com/cilium/cilium/pkg/logging/logfields"
 	"github.com/cilium/cilium/pkg/maps/callsmap"
 	"github.com/cilium/cilium/pkg/node/manager"
 	"github.com/cilium/cilium/pkg/option"
@@ -37,9 +30,6 @@ const (
 	subsystem = "datapath-loader"
 
 	symbolFromNetwork = "cil_from_network"
-
-	symbolToWireguard   = "cil_to_wireguard"
-	symbolFromWireguard = "cil_from_wireguard"
 
 	symbolFromHostNetdevXDP = "cil_xdp_entry"
 
@@ -109,76 +99,6 @@ func newLoader(p Params) *loader {
 		db:      p.DB,
 		devices: p.Devices,
 	}
-}
-
-func replaceWireguardDatapath(ctx context.Context, logger *slog.Logger, lnc *datapath.LocalNodeConfiguration, device netlink.Link) (err error) {
-	if err := compileWireguard(ctx, logger); err != nil {
-		return fmt.Errorf("compiling wireguard program: %w", err)
-	}
-
-	spec, err := ebpf.LoadCollectionSpec(wireguardObj)
-	if err != nil {
-		return fmt.Errorf("loading eBPF ELF %s: %w", wireguardObj, err)
-	}
-
-	cfg := config.NewBPFWireguard(config.NodeConfig(lnc))
-	cfg.InterfaceIfIndex = uint32(device.Attrs().Index)
-
-	cfg.EnableExtendedIPProtocols = option.Config.EnableExtendedIPProtocols
-	cfg.EnableNetkit = option.Config.DatapathMode == datapathOption.DatapathModeNetkit ||
-		option.Config.DatapathMode == datapathOption.DatapathModeNetkitL2
-	cfg.EphemeralMin = lnc.EphemeralMin
-
-	var obj wireguardObjects
-	commit, err := bpf.LoadAndAssign(logger, &obj, spec, &bpf.CollectionOptions{
-		Constants: cfg,
-		MapRenames: map[string]string{
-			"cilium_calls": fmt.Sprintf("cilium_calls_wireguard_%d", device.Attrs().Index),
-		},
-		CollectionOptions: ebpf.CollectionOptions{
-			Maps: ebpf.MapOptions{PinPath: bpf.TCGlobalsPath()},
-		},
-	})
-	if err != nil {
-		return err
-	}
-	defer obj.Close()
-
-	linkDir := bpffsDeviceLinksDir(bpf.CiliumPath(), device)
-	// Attach/detach cil_to_wireguard to/from egress.
-	if option.Config.NeedEgressOnWireGuardDevice(lnc.KPRConfig, lnc.EnableWireguard) {
-		if err := attachSKBProgram(logger, device, obj.ToWireguard, symbolToWireguard,
-			linkDir, netlink.HANDLE_MIN_EGRESS, option.Config.EnableTCX); err != nil {
-			return fmt.Errorf("interface %s egress: %w", device, err)
-		}
-	} else {
-		if err := detachSKBProgram(logger, device, symbolToWireguard,
-			linkDir, netlink.HANDLE_MIN_EGRESS); err != nil {
-			logger.Error("",
-				logfields.Error, err,
-				logfields.Device, device,
-			)
-		}
-	}
-	// Attach/detach cil_from_wireguard to/from ingress.
-	if option.Config.NeedIngressOnWireGuardDevice(lnc.KPRConfig, lnc.EnableWireguard) {
-		if err := attachSKBProgram(logger, device, obj.FromWireguard, symbolFromWireguard,
-			linkDir, netlink.HANDLE_MIN_INGRESS, option.Config.EnableTCX); err != nil {
-			return fmt.Errorf("interface %s ingress: %w", device, err)
-		}
-	} else {
-		if err := detachSKBProgram(logger, device, symbolFromWireguard,
-			linkDir, netlink.HANDLE_MIN_INGRESS); err != nil {
-			logger.Error("",
-				logfields.Error, err,
-				logfields.Device, device,
-			)
-		}
-	}
-	if err := commit(); err != nil {
-		return fmt.Errorf("committing bpf pins: %w", err)
-	}
-	return nil
 }
 
 // CallsMapPath gets the BPF Calls Map for the endpoint with the specified ID.
